+++ lib/runners/harness.rb
@@ -32,14 +32,7 @@ module Runners
       ensure_result do
         workspace = Workspace.prepare(options: options, working_dir: working_dir, trace_writer: trace_writer)
         workspace.open do |git_ssh_path, changes|
-          @config = conf = Config.load_from_dir(workspace.working_dir)
-
-          unless conf.ignore_patterns.empty?
-            trace_writer.message "Deleting ignored files..." do
-              files = Ignoring.new(workspace: workspace, ignore_patterns: conf.ignore_patterns).delete_ignored_files!
-              trace_writer.message(files.empty? ? "No deleted files." : "Successfully deleted #{files.size} file(s)")
-            end
-          end
+          @config = conf = workspace.config
 
           begin
             processor = processor_class.new(guid: guid, working_dir: working_dir, config: conf,