+++ lib/runners/workspace/git.rb
@@ -1,8 +1,10 @@
 module Runners
   class Workspace::Git < Workspace
     class Error < SystemError; end
+    class CloneFailed < Error; end
     class FetchFailed < Error; end
     class CheckoutFailed < Error; end
+    class SparseCheckoutFailed < Error; end
     class BlameFailed < Error; end
 
     def range_git_blame_info(path_string, start_line, end_line, trace: false)
@@ -21,24 +23,33 @@ module Runners
     end
 
     def prepare_head_source
-      shell.capture3!("git", "init", "--initial-branch=main")
-      shell.capture3!("git", "config", "gc.auto", "0")
-      shell.capture3!("git", "config", "advice.detachedHead", "false")
-      shell.capture3!("git", "config", "core.quotePath", "false")
-      shell.capture3!("git", "config", "core.hooksPath", mktmpdir.to_path) # NOTE: Prevent evil hooks from being executed.
-      shell.capture3!("git", "remote", "add", "origin", remote_url)
+      git_clone
+      git_setup
 
-      begin
-        shell.capture3_with_retry!("git", "fetch", *git_fetch_args, tries: try_count, sleep: sleep_lambda)
-      rescue Shell::ExecError => exn
-        raise FetchFailed, "git-fetch failed: #{exn.stderr_str}"
-      end
+      # Fetch a pull request if specified (including a forked repository).
+      git_fetch(git_source.refspec) unless git_source.refspec.empty?
 
-      begin
-        shell.capture3_with_retry!("git", "checkout", git_source.head, tries: try_count)
-      rescue Shell::ExecError => exn
-        raise CheckoutFailed, "git-checkout failed: #{exn.stderr_str}"
-      end
+      # First, fetch only a configuration file.
+      git_sparse_checkout_set "/#{Config::FILE_NAME}", "/#{Config::FILE_NAME_OLD}"
+      git_checkout
+
+      # Next, fetch remaining files except for *ignored* files.
+      git_sparse_checkout_set "/**", *config.ignore_patterns.map { |pat| "!#{pat}" }
+      git_checkout
+    end
+
+    def patches
+      return @patches if defined? @patches
+
+      base = git_source.base
+      head = git_source.head
+
+      @patches =
+        if base && head
+          # NOTE: We should use `...` (triple-dot) instead of `..` (double-dot). See https://git-scm.com/docs/git-diff
+          stdout, _ = shell.capture3!("git", "diff", "#{base}...#{head}", trace_stdout: false)
+          GitDiffParser.parse(stdout)
+        end
     end
 
     private
@@ -65,29 +76,49 @@ module Runners
       end
     end
 
-    def git_fetch_args
-      @git_fetch_args ||= [
-        '--quiet',
-        '--no-tags',
-        '--no-recurse-submodules',
-        'origin',
-        '+refs/heads/*:refs/remotes/origin/*',
-        *git_source.refspec,
-      ].freeze
+    # @see https://git-scm.com/docs/git-config
+    def git_setup
+      shell.capture3!("git", "config", "gc.auto", "0")
+      shell.capture3!("git", "config", "advice.detachedHead", "false")
+      shell.capture3!("git", "config", "core.quotePath", "false")
+      shell.capture3!("git", "config", "core.hooksPath", mktmpdir.to_path) # NOTE: Prevent evil hooks from being executed.
     end
 
-    def patches
-      return @patches if defined? @patches
+    # @see https://git-scm.com/docs/git-clone
+    # @see https://git-scm.com/docs/partial-clone
+    def git_clone
+      options = %w[
+        --filter=blob:none
+        --no-checkout
+        --no-recurse-submodules
+        --no-tags
+        --quiet
+      ]
+      shell.capture3_with_retry!("git", "clone", *options, "--", remote_url, ".", tries: try_count, sleep: sleep_lambda)
+    rescue Shell::ExecError => exn
+      raise CloneFailed, "git-clone failed: #{exn.stderr_str}"
+    end
 
-      base = git_source.base
-      head = git_source.head
+    # @see https://git-scm.com/docs/git-fetch
+    def git_fetch(refspec)
+      shell.capture3_with_retry!("git", "fetch", "origin", *refspec, tries: try_count, sleep: sleep_lambda)
+    rescue Shell::ExecError => exn
+      raise FetchFailed, "git-fetch failed: #{exn.stderr_str}"
+    end
 
-      @patches =
-        if base && head
-          # NOTE: We should use `...` (triple-dot) instead of `..` (double-dot). See https://git-scm.com/docs/git-diff
-          stdout, _ = shell.capture3!("git", "diff", "#{base}...#{head}", trace_stdout: false)
-          GitDiffParser.parse(stdout)
-        end
+    # @see https://git-scm.com/docs/git-checkout
+    def git_checkout
+      shell.capture3_with_retry!("git", "checkout", git_source.head, tries: try_count)
+    rescue Shell::ExecError => exn
+      raise CheckoutFailed, "git-checkout failed: #{exn.stderr_str}"
+    end
+
+    # @see https://git-scm.com/docs/git-sparse-checkout
+    # @see https://git-scm.com/docs/gitignore
+    def git_sparse_checkout_set(*patterns)
+      shell.capture3!("git", "sparse-checkout", "set", *patterns)
+    rescue Shell::ExecError => exn
+      raise SparseCheckoutFailed, "git-sparse-checkout failed: #{exn.stderr_str}"
     end
   end
 end